requirements.txt
pillow
pypresence
colorama